requirements.txt
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Faker==4.0.2
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
httpie==2.0.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.6.3
MarkupSafe==1.1.1
pathtools==0.1.2
Pygments==2.6.1
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
selenium==3.141.0
six==1.14.0
spyne==2.13.16
SQLAlchemy==1.3.15
suds-py3==1.4.4.1
text-unidecode==1.3
urllib3==1.25.8
watchdog==2.0.3
Werkzeug==1.0.1
WTForms==2.2.1
